// src/types/woocommerce/orders/orderstatus.ts
export enum WPOrderStatus {
    Pending = 'wc-pending',
    Processing = 'wc-processing',
    OnHold = 'wc-on-hold',
    Completed = 'wc-completed',
    Cancelled = 'wc-cancelled',  
    Failed = 'wc-failed',
    Trash = 'trash',
    AutoDraft = 'auto-draft',
    NewOrder = 'wc-new-order',
    Shipping = 'wc-shipping',
}